# Conflicts:
#	src/action/molecule.hpp
#	src/view/ui/widget/molecule_scene_view.cpp
#	src/view/ui/widget/molecule_scene_view.hpp

+++ src/action/action_manager.cpp
@@ -167,7 +167,7 @@ namespace VTX
 			_bufferRedo.pop_front();
 		}
 
-		void ActionManager::update( const double & p_deltaTime )
+		void ActionManager::update( const float & p_deltaTime )
 		{
 			while ( _actionQueue.empty() == false )
 			{
