+++ src/model/path.cpp
@@ -14,10 +14,10 @@ namespace VTX
 		Path::Path() : BaseModel( ID::Model::MODEL_PATH ) {};
 		Path::~Path() { MVC::MvcManager::get().deleteAllModels( _viewpoints ); }
 
-		const std::vector<std::string> * const Path::getCurrentActions( const double p_time )
+		const std::vector<std::string> * const Path::getCurrentActions( const float p_time )
 		{
 			uint  size	 = (uint)_viewpoints.size();
-			float total	 = 0.0;
+			float total	 = 0.f;
 			uint  offset = 0;
 
 			// Get the last actions.
@@ -35,13 +35,13 @@ namespace VTX
 			return &_viewpoints[ offset - 1 ]->getActions();
 		}
 
-		Model::Viewpoint Path::getInterpolatedViewpoint( const double p_time ) const
+		Model::Viewpoint Path::getInterpolatedViewpoint( const float p_time ) const
 		{
 			Viewpoint viewpoint = Viewpoint( (Path * const)this );
 
-			uint   size	  = (uint)_viewpoints.size();
-			double total  = 0.0;
-			uint   offset = 0;
+			uint  size	 = (uint)_viewpoints.size();
+			float total	 = 0.f;
+			uint  offset = 0;
 
 			// Find the next and previous points.
 			while ( total <= p_time && offset < size )
@@ -55,7 +55,7 @@ namespace VTX
 				// Computes value.
 				Viewpoint * const p0	= _viewpoints[ offset - 1 ];
 				Viewpoint * const p1	= _viewpoints[ Util::Math::min<int>( (int)size - 1, offset ) ];
-				double			  value = 1.0 - ( ( total - p_time ) / p1->getDuration() );
+				float			  value = 1.f - ( ( total - p_time ) / p1->getDuration() );
 
 				// Lerp.
 				viewpoint.setPosition( Util::Math::linearInterpolation( p0->getPosition(), p1->getPosition(), value ) );
@@ -71,7 +71,7 @@ namespace VTX
 				Viewpoint * const p1	= _viewpoints[ offset - 1 ];
 				Viewpoint * const p2	= _viewpoints[ Util::Math::min<int>( (int)size - 1, offset ) ];
 				Viewpoint * const p3	= _viewpoints[ Util::Math::min<int>( (int)size - 1, offset + 1 ) ];
-				double			  value = 1.0 - ( ( total - p_time ) / p2->getDuration() );
+				float			  value = 1.f - ( ( total - p_time ) / p2->getDuration() );
 
 				viewpoint.setPosition( Util::Math::catmullRomInterpolation( p0->getPosition(), p1->getPosition(), p2->getPosition(), p3->getPosition(), value ) );
 				viewpoint.setRotation( Util::Math::catmullRomInterpolation( p0->getRotation(), p1->getRotation(), p2->getRotation(), p3->getRotation(), value ) );
@@ -88,7 +88,7 @@ namespace VTX
 				Viewpoint * const p1	= _viewpoints[ offset - 1 ];
 				Viewpoint * const p2	= _viewpoints[ Util::Math::min<int>( (int)size - 1, offset ) ];
 				Viewpoint * const p3	= _viewpoints[ Util::Math::min<int>( (int)size - 1, offset + 1 ) ];
-				double			  value = 1.0 - ( ( total - p_time ) / p2->getDuration() );
+				float			  value = 1.f - ( ( total - p_time ) / p2->getDuration() );
 
 				viewpoint.setPosition( Util::Math::cubicInterpolation( p0->getPosition(), p1->getPosition(), p2->getPosition(), p3->getPosition(), value ) );
 				viewpoint.setRotation( Util::Math::cubicInterpolation( p0->getRotation(), p1->getRotation(), p2->getRotation(), p3->getRotation(), value ) );
@@ -117,7 +117,7 @@ namespace VTX
 				float duration = 0.f;
 				if ( _viewpoints.size() >= 2 )
 				{
-					duration = _duration / (float)( _viewpoints.size() - 1 );
+					duration = _duration / ( _viewpoints.size() - 1 );
 				}
 
 				for ( uint i = 1; i < _viewpoints.size(); ++i )
@@ -144,7 +144,7 @@ namespace VTX
 				uint  size			= (uint)Util::Math::max<int>( (int)_viewpoints.size() - 1, 0 );
 				for ( uint i = 0; i < size; ++i )
 				{
-					totalDistance += (float)Util::Math::distance( _viewpoints[ i ]->getPosition(), _viewpoints[ i + 1u ]->getPosition() );
+					totalDistance += Util::Math::distance( _viewpoints[ i ]->getPosition(), _viewpoints[ i + 1u ]->getPosition() );
 				}
 
 				// Compute viewpoint durations.
@@ -156,7 +156,7 @@ namespace VTX
 						viewpoint->setDuration( 0.f );
 						break;
 					}
-					float distance = (float)Util::Math::distance( _viewpoints[ i - 1u ]->getPosition(), viewpoint->getPosition() );
+					float distance = Util::Math::distance( _viewpoints[ i - 1u ]->getPosition(), viewpoint->getPosition() );
 					viewpoint->setDuration( _duration * distance / totalDistance );
 				}
 			}
