+++ src/state/state_machine.cpp
@@ -48,7 +48,7 @@ namespace VTX
 			}
 		}
 
-		void StateMachine::update( const double & p_deltaTime )
+		void StateMachine::update( const float & p_deltaTime )
 		{
 			if ( _currentState != nullptr )
 			{
